backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
bcrypt
google-generativeai
python-dotenv
email-validator
